requirements.txt
Faker==4.0.2
google-cloud-firestore==1.6.2
